refactor Operation factory to specialised factories
